+++ packages/react-article-components/src/components/img-with-placeholder/index.js
@@ -179,12 +179,26 @@ export default class Img extends React.PureComponent {
     }, 1500)
   }
 
+  _renderImagePlaceholder() {
+    const { toShowPlaceholder } = this.state
+    const { imgPlaceholderSrc } = this.props
+    return imgPlaceholderSrc ? (
+      <ImgPlaceholder
+        src={replaceGCSUrlOrigin(imgPlaceholderSrc)}
+        toShow={toShowPlaceholder}
+      />
+    ) : (
+      <Placeholder toShow={toShowPlaceholder}>
+        <PlaceholderIcon />
+      </Placeholder>
+    )
+  }
+
   render() {
-    const { isLoaded, toShowPlaceholder } = this.state
+    const { isLoaded } = this.state
     const {
       alt,
       className,
-      imgPlaceholderSrc,
       imgProps,
       imageSet,
       defaultImage,
@@ -193,16 +207,28 @@ export default class Img extends React.PureComponent {
       sizes,
     } = this.props
 
+    const defaultImageOriginalUrl = _.get(defaultImage, 'url')
+    /* Render placeholder only if no valid image is given */
+    if (!defaultImageOriginalUrl && (!imageSet || imageSet.length === 0)) {
+      return (
+        <ImgContainer className={className} heightString="height: 100%;">
+          {this._renderImagePlaceholder()}
+        </ImgContainer>
+      )
+    }
+
     const srcset = getSrcsetString(imageSet)
+    const isObjectFit = Boolean(objectFit)
     const heightWidthRatio =
       _.get(defaultImage, 'height') / _.get(defaultImage, 'width')
-    if (!heightWidthRatio) {
+    if (isObjectFit && !heightWidthRatio) {
       console.warn(
-        'Valid height and width of the default image are required. But the default image is:',
+        'Warning on Img component:',
+        'The `objecFit` is set, but no valid height/width ratio of `props.defaultImage` is given.',
+        '`props.defaultImage`:',
         defaultImage
       )
     }
-    const isObjectFit = Boolean(objectFit)
     const defaultImageSrc = replaceGCSUrlOrigin(_.get(defaultImage, 'url'))
 
     return (
@@ -214,16 +240,7 @@ export default class Img extends React.PureComponent {
             : `padding-top: ${heightWidthRatio * 100}%;`
         }
       >
-        {imgPlaceholderSrc ? (
-          <ImgPlaceholder
-            src={replaceGCSUrlOrigin(imgPlaceholderSrc)}
-            toShow={toShowPlaceholder}
-          />
-        ) : (
-          <Placeholder toShow={toShowPlaceholder}>
-            <PlaceholderIcon />
-          </Placeholder>
-        )}
+        {this._renderImagePlaceholder()}
         <ImgBox toShow={isLoaded}>
           {isObjectFit ? (
             <React.Fragment>