+++ packages/react-article-components/src/constants/element-alignment.js
@@ -1,0 +1,6 @@
+export default {
+  left: 'left',
+  center: 'center',
+  right: 'right',
+  centerSmall: 'center-small',
+}