+++ packages/react-article-components/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## RELEASE
 
+### 1.0.13
+
+- Render placeholder if there’s no any image given to img-with-placeholder
+- Update warning of image-with-placeholder
+- Add `center-small` image
+
 ### 1.0.12
 
 - Add margin between topic and title of leading if there's no subtitle. By changing margin-top of subtitle to the margin-bottom of topic.